refactor recognition services for web
